Entropy

Calculates entropy based on heights, 2D, or 3D positions of points.

Usage

entropy(data, ...)

Arguments

data

Digital elevation model of class RasterLayer, a triangular mesh of class mesh3d, or a pointcloud of class data.frame.

...

Additional arguments depending on the data type.

Value

A value for entropy.

Details

If data is a 3D pointcloud or 3D mesh, entropy_3d() function will be applied; if data is a 2D pointcloud,entropy_2d() will be applied; and if data is a RasterLayer, entropy_1d() will be applied after conversion to a vector. See the documentation of those functions for details on the necessary arguments.

References

X. Liu, Q. Ma, X. Wu, T. Hu, Z. Liu, L. Liu, Q. Guo, Y. Su (2022). A novel entropy-based method to quantify forest canopy structural complexity from multiplatform lidar point clouds. Remote Sens. Environ. 282, 113280.

See also

Examples

entropy(mcap, bw = 0.02, grid_size = 0.01)
#> [1] 11.21657
entropy(horseshoe, grid_size = 0.05)
#> [1] 3.118644