Height Variation in cells at different scales

Height Variation in cells at different scales

hvar(
  data,
  x,
  y,
  lvec,
  L,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)

Arguments

data

Digital elevation model of class RasterLayer.

x

Bottom-left of bounding box.

y

Bottom-left of bounding box.

lvec

scales to use for calculation

L

Bounding box extent (i.e., side length).

parallel

TRUE or FALSE. Use parallel processing? Note: parallel must be installed.

ncores

number of cores to use when parallel = TRUE.

Value

A data.frame containing height ranges of cells at different scales.

Details

This function is used for calculating fractal dimension using the height variation method.

Examples

hvar(horseshoe, x = -470, y = 1266, L = 2, lvec = c(0.125, 0.25, 0.5, 1))
#>         l           h
#> 1   0.125 0.153405867
#> 2   0.125 0.300532856
#> 3   0.125 0.308145838
#> 4   0.125 0.301096802
#> 5   0.125 0.285108004
#> 6   0.125 0.115060081
#> 7   0.125 0.236937141
#> 8   0.125 0.259187708
#> 9   0.125 0.393820066
#> 10  0.125 0.194270954
#> 11  0.125 0.153722563
#> 12  0.125 0.140229340
#> 13  0.125 0.059330740
#> 14  0.125 0.057614260
#> 15  0.125 0.028143377
#> 16  0.125 0.021279154
#> 17  0.125 0.033430290
#> 18  0.125 0.119361200
#> 19  0.125 0.171767502
#> 20  0.125 0.061070156
#> 21  0.125 0.089295778
#> 22  0.125 0.147761393
#> 23  0.125 0.069798870
#> 24  0.125 0.092623281
#> 25  0.125 0.332383356
#> 26  0.125 0.230841713
#> 27  0.125 0.163960152
#> 28  0.125 0.127236214
#> 29  0.125 0.081798811
#> 30  0.125 0.053572750
#> 31  0.125 0.058292923
#> 32  0.125 0.068348818
#> 33  0.125 0.055659018
#> 34  0.125 0.023751764
#> 35  0.125 0.191447620
#> 36  0.125 0.069222946
#> 37  0.125 0.104235420
#> 38  0.125 0.062363062
#> 39  0.125 0.055037584
#> 40  0.125 0.053127365
#> 41  0.125 0.152286854
#> 42  0.125 0.440485182
#> 43  0.125 0.163979883
#> 44  0.125 0.134149628
#> 45  0.125 0.092401342
#> 46  0.125 0.038307638
#> 47  0.125 0.048501406
#> 48  0.125 0.062243338
#> 49  0.125 0.036411486
#> 50  0.125 0.018131208
#> 51  0.125 0.013504572
#> 52  0.125 0.458781586
#> 53  0.125 0.641067238
#> 54  0.125 0.238799171
#> 55  0.125 0.123776093
#> 56  0.125 0.197380342
#> 57  0.125 0.063220301
#> 58  0.125 0.318067646
#> 59  0.125 0.234170160
#> 60  0.125 0.126290398
#> 61  0.125 0.106782684
#> 62  0.125 0.046310596
#> 63  0.125 0.041709671
#> 64  0.125 0.044047213
#> 65  0.125 0.031075583
#> 66  0.125 0.021373043
#> 67  0.125 0.015366755
#> 68  0.125 0.015480766
#> 69  0.125 0.236433487
#> 70  0.125 0.391416521
#> 71  0.125 0.566834812
#> 72  0.125 0.544762402
#> 73  0.125 0.259054613
#> 74  0.125 0.112234983
#> 75  0.125 0.349158859
#> 76  0.125 0.150525751
#> 77  0.125 0.080518265
#> 78  0.125 0.077973108
#> 79  0.125 0.022172956
#> 80  0.125 0.031638851
#> 81  0.125 0.078251238
#> 82  0.125 0.087751122
#> 83  0.125 0.083402596
#> 84  0.125 0.045724201
#> 85  0.125 0.014000778
#> 86  0.125 0.130339622
#> 87  0.125 0.173221035
#> 88  0.125 0.083436756
#> 89  0.125 0.444054794
#> 90  0.125 0.587276726
#> 91  0.125 0.266940851
#> 92  0.125 0.337076950
#> 93  0.125 0.112714367
#> 94  0.125 0.084664688
#> 95  0.125 0.070355577
#> 96  0.125 0.033725176
#> 97  0.125 0.034886217
#> 98  0.125 0.031830559
#> 99  0.125 0.101721935
#> 100 0.125 0.069963455
#> 101 0.125 0.025193853
#> 102 0.125 0.033310928
#> 103 0.125 0.139929447
#> 104 0.125 0.121944437
#> 105 0.125 0.106463718
#> 106 0.125 0.156219425
#> 107 0.125 0.570140448
#> 108 0.125 0.124132519
#> 109 0.125 0.261015406
#> 110 0.125 0.199652815
#> 111 0.125 0.156811504
#> 112 0.125 0.073110476
#> 113 0.125 0.035925007
#> 114 0.125 0.059299984
#> 115 0.125 0.068809156
#> 116 0.125 0.043066750
#> 117 0.125 0.047123013
#> 118 0.125 0.009291868
#> 119 0.125 0.010307732
#> 120 0.125 0.113728733
#> 121 0.125 0.242895603
#> 122 0.125 0.566553144
#> 123 0.125 0.536162415
#> 124 0.125 0.053242168
#> 125 0.125 0.072932043
#> 126 0.125 0.557369385
#> 127 0.125 0.207315731
#> 128 0.125 0.099931917
#> 129 0.125 0.084100389
#> 130 0.125 0.057709866
#> 131 0.125 0.026297770
#> 132 0.125 0.053634834
#> 133 0.125 0.033627081
#> 134 0.125 0.053793411
#> 135 0.125 0.021264601
#> 136 0.125 0.008598061
#> 137 0.125 0.297683592
#> 138 0.125 0.569891930
#> 139 0.125 0.363071089
#> 140 0.125 0.059033012
#> 141 0.125 0.091350079
#> 142 0.125 0.083690233
#> 143 0.125 0.567886343
#> 144 0.125 0.149549990
#> 145 0.125 0.112108994
#> 146 0.125 0.090015125
#> 147 0.125 0.077953053
#> 148 0.125 0.077643766
#> 149 0.125 0.070821123
#> 150 0.125 0.040020971
#> 151 0.125 0.044977770
#> 152 0.125 0.035824299
#> 153 0.125 0.011575327
#> 154 0.125 0.339800510
#> 155 0.125 0.047929459
#> 156 0.125 0.147299767
#> 157 0.125 0.165950851
#> 158 0.125 0.100343180
#> 159 0.125 0.060374479
#> 160 0.125 0.715530014
#> 161 0.125 0.290083838
#> 162 0.125 0.134966211
#> 163 0.125 0.108396921
#> 164 0.125 0.094444675
#> 165 0.125 0.107640314
#> 166 0.125 0.033785582
#> 167 0.125 0.038255234
#> 168 0.125 0.046247463
#> 169 0.125 0.037146721
#> 170 0.125 0.017923555
#> 171 0.125 0.126311741
#> 172 0.125 0.047275620
#> 173 0.125 0.112006931
#> 174 0.125 0.160969172
#> 175 0.125 0.065505104
#> 176 0.125 0.057969513
#> 177 0.125 0.275360479
#> 178 0.125 0.554142466
#> 179 0.125 0.144457836
#> 180 0.125 0.112244873
#> 181 0.125 0.113566923
#> 182 0.125 0.076450415
#> 183 0.125 0.037904158
#> 184 0.125 0.068495054
#> 185 0.125 0.060243835
#> 186 0.125 0.034979458
#> 187 0.125 0.042234735
#> 188 0.125 0.076208572
#> 189 0.125 0.080751038
#> 190 0.125 0.116474705
#> 191 0.125 0.137028418
#> 192 0.125 0.102404957
#> 193 0.125 0.126828489
#> 194 0.125 0.134408607
#> 195 0.125 0.565452900
#> 196 0.125 0.256641731
#> 197 0.125 0.143483753
#> 198 0.125 0.144562263
#> 199 0.125 0.114528666
#> 200 0.125 0.078411388
#> 201 0.125 0.057088013
#> 202 0.125 0.051049805
#> 203 0.125 0.063772898
#> 204 0.125 0.043835011
#> 205 0.125 0.056595526
#> 206 0.125 0.058197937
#> 207 0.125 0.052795467
#> 208 0.125 0.074639034
#> 209 0.125 0.054155750
#> 210 0.125 0.108632174
#> 211 0.125 0.128687267
#> 212 0.125 0.267371311
#> 213 0.125 0.365769567
#> 214 0.125 0.437550001
#> 215 0.125 0.275106668
#> 216 0.125 0.146726723
#> 217 0.125 0.127928658
#> 218 0.125 0.110360069
#> 219 0.125 0.099457979
#> 220 0.125 0.128642044
#> 221 0.125 0.151002483
#> 222 0.125 0.063744078
#> 223 0.125 0.046342487
#> 224 0.125 0.048592224
#> 225 0.125 0.094460449
#> 226 0.125 0.137602873
#> 227 0.125 0.228856173
#> 228 0.125 0.093498945
#> 229 0.125 0.070544100
#> 230 0.125 0.043134050
#> 231 0.125 0.173572798
#> 232 0.125 0.618198433
#> 233 0.125 0.320903568
#> 234 0.125 0.173800945
#> 235 0.125 0.109796162
#> 236 0.125 0.101427155
#> 237 0.125 0.106795683
#> 238 0.125 0.090318069
#> 239 0.125 0.116921415
#> 240 0.125 0.081090870
#> 241 0.125 0.114550066
#> 242 0.125 0.181513548
#> 243 0.125 0.101551476
#> 244 0.125 0.122265148
#> 245 0.125 0.140113459
#> 246 0.125 0.043745556
#> 247 0.125 0.091548891
#> 248 0.125 0.462540874
#> 249 0.125 0.587186060
#> 250 0.125 0.642839022
#> 251 0.125 0.247268476
#> 252 0.125 0.100296412
#> 253 0.125 0.075356760
#> 254 0.125 0.105126953
#> 255 0.125 0.089692850
#> 256 0.125 0.086200495
#> 257 0.125 0.078385973
#> 258 0.125 0.175146513
#> 259 0.125 0.101030569
#> 260 0.125 0.084087257
#> 261 0.125 0.087432680
#> 262 0.125 0.123794231
#> 263 0.125 0.135814695
#> 264 0.125 0.441878119
#> 265 0.125 0.582480793
#> 266 0.125 0.352582846
#> 267 0.125 0.058265066
#> 268 0.125 0.308131485
#> 269 0.125 0.153404827
#> 270 0.125 0.130831614
#> 271 0.125 0.195430622
#> 272 0.125 0.134792414
#> 273 0.125 0.059985905
#> 274 0.125 0.034435492
#> 275 0.125 0.155754595
#> 276 0.125 0.036833181
#> 277 0.125 0.175928307
#> 278 0.125 0.128956680
#> 279 0.125 0.101971760
#> 280 0.125 0.056796255
#> 281 0.125 0.398428192
#> 282 0.125 0.243039055
#> 283 0.125 0.024862461
#> 284 0.125 0.124383307
#> 285 0.125 0.333305483
#> 286 0.125 0.152124386
#> 287 0.125 0.146425896
#> 288 0.125 0.304217691
#> 289 0.125 0.231933756
#> 290 0.250 0.311159945
#> 291 0.250 0.374192619
#> 292 0.250 0.317052755
#> 293 0.250 0.558002043
#> 294 0.250 0.447078199
#> 295 0.250 0.235819378
#> 296 0.250 0.071625023
#> 297 0.250 0.081833124
#> 298 0.250 0.663852911
#> 299 0.250 0.370063505
#> 300 0.250 0.232877846
#> 301 0.250 0.554987383
#> 302 0.250 0.262541742
#> 303 0.250 0.058642588
#> 304 0.250 0.097294931
#> 305 0.250 0.039845047
#> 306 0.250 0.518064032
#> 307 0.250 0.678709583
#> 308 0.250 0.503139849
#> 309 0.250 0.473762083
#> 310 0.250 0.195383101
#> 311 0.250 0.065377083
#> 312 0.250 0.106315002
#> 313 0.250 0.059006634
#> 314 0.250 0.472406120
#> 315 0.250 0.702872028
#> 316 0.250 0.604946136
#> 317 0.250 0.607550125
#> 318 0.250 0.233680601
#> 319 0.250 0.071019478
#> 320 0.250 0.082471066
#> 321 0.250 0.039306726
#> 322 0.250 0.534459572
#> 323 0.250 0.210413828
#> 324 0.250 0.140495129
#> 325 0.250 0.748603249
#> 326 0.250 0.257985344
#> 327 0.250 0.161145945
#> 328 0.250 0.068861399
#> 329 0.250 0.070519810
#> 330 0.250 0.115587244
#> 331 0.250 0.170550032
#> 332 0.250 0.142218285
#> 333 0.250 0.748567438
#> 334 0.250 0.489763927
#> 335 0.250 0.261766491
#> 336 0.250 0.126742764
#> 337 0.250 0.168605785
#> 338 0.250 0.114559898
#> 339 0.250 0.189748564
#> 340 0.250 0.314628897
#> 341 0.250 0.198655396
#> 342 0.250 0.659192600
#> 343 0.250 0.824803219
#> 344 0.250 0.263203440
#> 345 0.250 0.270072889
#> 346 0.250 0.154436750
#> 347 0.250 0.217546320
#> 348 0.250 0.204115925
#> 349 0.250 0.377811928
#> 350 0.250 0.770802097
#> 351 0.250 0.814877968
#> 352 0.250 0.336207399
#> 353 0.250 0.529216547
#> 354 0.500 0.698796701
#> 355 0.500 0.706519556
#> 356 0.500 0.544635677
#> 357 0.500 0.125044279
#> 358 0.500 0.972217703
#> 359 0.500 0.828354168
#> 360 0.500 0.258774185
#> 361 0.500 0.118645573
#> 362 0.500 0.601266451
#> 363 0.500 0.810004463
#> 364 0.500 0.662769184
#> 365 0.500 0.170902138
#> 366 0.500 0.290751572
#> 367 0.500 0.517153606
#> 368 0.500 0.893439550
#> 369 0.500 0.741272621
#> 370 1.000 1.044014063
#> 371 1.000 0.606689758
#> 372 1.000 0.810004463
#> 373 1.000 1.048795605