Geometric plane equation

Calculates either rugosity, fractal dimension or height range based on the other two variables.

rdh_theory(r, d, h, L, L0)

Arguments

r, d, h

two of the three variables to calculate the third.

L

extent.

L0

resolution.

Value

A value corresponding one of the three variables not given to the function.

Details

This function uses the geometric plane equation from Torres-Pulliza et al. (2020) to calculate one of rugosity, fractal dimension or height range based on the other two variables.

References

Torres-Pulliza D, Dornelas M, Pizarro O, Bewley M, Blowes SA, Boutros N, Brambilla V, Chase TJ, Frank G, Friedman A, Hoogenboom MO, Williams S, Zawada KJA, Madin JS (2020) A geometric basis for surface habitat complexity and biodiversity. Nature Ecology & Evolution 4:1495-1501. https://doi.org/10.1038/s41559-020-1281-8

Examples

rdh_theory(r=4, h=1, L=1, L0=0.01)
#> $d
#> [1] 2.36928
#> 
rdh_theory(d=2.36928, h=1, L=1, L0=0.01)
#> $r
#> [1] 3.999995
#> 
rdh_theory(d=2.36928, r=4, L=1, L0=0.01)
#> $h
#> [1] 1.000001
#>