Height Variation in cells at different scales

Height Variation in cells at different scales

hvar(
  dem,
  x,
  y,
  L0,
  L,
  n = 5,
  parallel = FALSE,
  ncores = (parallel::detectCores() - 1)
)

Arguments

dem

Digital elevation model in raster format.

x

Bottom-left of bounding box.

y

Bottom-left of bounding box.

L0

Resolution

L

Bounding box extent (i.e., side length).

n

number of scales to use

parallel

TRUE or FALSE. Use parallel computation?

ncores

number of cores to use when parallel = TRUE.

Value

A data.frame containing height ranges of cells at different scales.

Details

This function is used for calculating fractal dimension using the height variation method. It is separated from the fd function because it takes a while to run, and so it is useful to save the output before using the fd function.

Examples

hvar(horseshoe, x = -470, y = 1266, L = 2, L0 = 0.5)
#>           L0        H0
#> 1  0.5000000 0.2930083
#> 2  0.5000000 0.5216577
#> 3  0.5000000 0.9197886
#> 4  0.5000000 0.7609899
#> 5  0.5000000 0.6210294
#> 6  0.5000000 0.8130283
#> 7  0.5000000 0.6640019
#> 8  0.5000000 0.1832321
#> 9  0.5000000 0.9781842
#> 10 0.5000000 0.8325553
#> 11 0.5000000 0.2785048
#> 12 0.5000000 0.1432183
#> 13 0.5000000 0.7134392
#> 14 0.5000000 0.7151203
#> 15 0.5000000 0.6334853
#> 16 0.5000000 0.1260359
#> 17 0.6597540 0.4046052
#> 18 0.6597540 0.8875833
#> 19 0.6597540 0.9153261
#> 20 0.6597540 1.0143380
#> 21 0.6597540 1.0150182
#> 22 0.6597540 0.1954219
#> 23 0.6597540 0.9852901
#> 24 0.6597540 0.9052749
#> 25 0.6597540 0.2543252
#> 26 0.8705506 0.8130283
#> 27 0.8705506 1.0453691
#> 28 0.8705506 1.0463586
#> 29 0.8705506 0.7587984
#> 30 1.1486984 0.9395239
#> 31 1.5157166 1.0660069
#> 32 2.0000000 1.0917811